Cambios en el parser:
Pequeños bugs solucionados:
-Se validaba el ancho/alto en los circulos, y no se permitia que se
crearan
-Se estan devolviendo valores doubles para altoUn y anchoUn del
escenario

+++ tallerDeProgramacionI-SnowBros/headers/Parser/ParserValidator.h
@@ -55,8 +55,8 @@ private:
 	//Validaciones en el escenario
 	static int valAltoPx(Json::Value esc);
 	static int valAnchoPx(Json::Value esc);
-	static int valAltoUn(Json::Value esc);
-	static int valAnchoUn(Json::Value esc);
+	static double valAltoUn(Json::Value esc);
+	static double valAnchoUn(Json::Value esc);
 	static std::string valImagenFondo(Json::Value esc);
 
 	//Validaciones en el personaje
